requirements.txt
fastapi
uvicorn
python-multipart
fastapi
uvicorn
python-multipart
jinja2
pandas
pyyaml
python-whois
dnspython
cryptography
pyopenssl
requests
aiohttp
PyGithub
python-gitlab
python-telegram-bot
pydantic-settings
loguru
